+++ tests/mig_add_gm2mfield/models.py
@@ -2,7 +2,16 @@ from django.db import models
 
 from gm2m import GM2MField
 
-from ..app.models import Project, Task
+from ..app.models import Task
+
+
+# this is for bug #33, where a circular relation is needed
+class Project(models.Model):
+
+    class Meta:
+        app_label = 'mig_add_gm2mfield'
+
+    owner = models.ForeignKey('User')
 
 
 class User(models.Model):
@@ -11,4 +20,5 @@ class User(models.Model):
         app_label = 'mig_add_gm2mfield'
 
     name = models.CharField(blank=True, default='', max_length=100)
-    # items = GM2MField(Project, Task)
+    # mig1: items = GM2MField(Project)
+    # mig2: items = GM2MField(Task, Project)